temp commit

+++ packages/react/src/current_batch_config.rs
@@ -1,5 +1,5 @@
 pub struct ReactCurrentBatchConfig {
-    pub transition: u8,
+    pub transition: u32,
 }
 
 pub static mut REACT_CURRENT_BATCH_CONFIG: ReactCurrentBatchConfig =
